+++ test/fetch.spec.js
@@ -1,11 +1,11 @@
-describe('Client',  function () {
+describe('Client', function () {
   jasmine.DEFAULT_TIMEOUT_INTERVAL = 5000
 
   var sandboxId = 'Y1k3xBDc'
-  var protocols = ['http', 'https']
+  var protocols = ['http']
   var server = '://api.zpush.io/' + sandboxId
 
-  describe('Fetch API Url',  function () {
+  describe('Fetch API Url', function () {
     protocols.forEach(function (protocol) {
       it('Should get servers list on ' + protocol + ' sever', function (done) {
         var url = protocol + server
@@ -13,10 +13,12 @@ describe('Client',  function () {
           .then(function (response) {
             return response.json()
           })
-          .then(function (servers) {
+          .then(function (response) {
+            expect(response.servers).toBeTruthy()
             done()
           })
           .catch(function (error) {
+            expect(error).toBeNull()
             done()
           })
       })